A separated rendering context for all the document. Config is now the local rendering context of a renderer or a sub-renderer.
