+++ lib/Skriv/Markup/Renderer.php
@@ -10,23 +10,46 @@ namespace Skriv\Markup;
  * @package	SkrivMarkup
  * @see		WikiRenderer
  */
+use WikiRenderer\Config;
+
 class Renderer {
-	/** The configuration object. */
-	protected $_config = null;
+	/** @var RenderingContext */
+	private $_context;
+	/** @var Config */
+	private $_config;
 	/** The WikiRenderer object. */
-	protected $_wikiRenderer = null;
+	private $_wikiRenderer;
 
 	/**
 	 * Factory method. Creates a renderer object of the given type.
-	 * @param	string	$type	(optional) Type of rendering object. "html" by default.
+	 * @param	string	$type	(optional). Available values: 'html' (by default) and 'docbook' (warning, docbook is still in alpha version).
 	 * @param	array	$params	(optional) Hash of parameters. The accepted parameters depends of the chosen rendering type.
 	 * @return	\Skriv\Markup\Renderer	A rendering object.
-	 * @throws	Exception	If something goes wrong.
+	 * @throws	\Exception	If something goes wrong.
 	 */
-	static public function factory($type='html', array $params=null) {
-		if (!isset($type) || !strcasecmp($type, 'html'))
-			return (new Html\Renderer($params));
-		throw new \Exception("Unknown Skriv rendering type '$type'.");
+	static public function factory($type = 'html', array $params = null) {
+		if (!isset($type))
+			$type = 'html';
+		switch ($type) {
+			case 'html':
+				$context = new Html\HtmlRenderingContext($params);
+				return new Renderer($context, new Html\Config($context));
+			case 'docbook':
+				$context = new DocBook\DocBookRenderingContext($params);
+				return new Renderer($context, new DocBook\Config($context));
+			default:
+				throw new \Exception("Unknown Skriv rendering type '$type'.");
+		}
+	}
+	/**
+	 * Parses a Skriv text and generates a converted text.
+	 * @param	string	$text	The text to parse.
+	 * @return	string	The generated string.
+	 */
+	public function __construct(RenderingContext $context, Config $config) {
+		$this->_context = $context;
+		$this->_config = $config;
+		$this->_wikiRenderer = new \WikiRenderer\Renderer($config);
 	}
 	/**
 	 * Parses a Skriv text and generates a converted text.
@@ -34,16 +57,16 @@ class Renderer {
 	 * @return	string	The generated string.
 	 */
 	public function render($text) {
-		return ($this->_wikiRenderer->render($text));
+		return $this->_wikiRenderer->render($text);
 	}
 	/**
-         * Returns the TOC content. By default, the rendered string is returned, but the
-         * raw TOC tree is available.
-         * @param	bool	$raw	(optional) Set to True to get the raw TOC tree. False by default.
-         * @return	string|array	The TOC rendered string or the TOC tree.
-         */
-        public function getToc($raw=false) {
-		return ($this->_config->getToc($raw));
+	 * Returns the TOC content. By default, the rendered string is returned, but the
+	 * raw TOC tree is available.
+	 * @param	bool	$raw	(optional) Set to True to get the raw TOC tree. False by default.
+	 * @return	string|array	The TOC rendered string or the TOC tree.
+	 */
+	public function getToc($raw = false) {
+		return ($this->_context->getToc($raw));
 	}
 	/**
 	 * Returns the footnotes content. By default, the rendered string is returned, but the
@@ -52,8 +75,8 @@ class Renderer {
 	 *				False by default.
 	 * @return	string|array	The footnotes' rendered string or the list of footnotes.
 	 */
-	public function getFootnotes($raw=false) {
-		return ($this->_config->getFootnotes($raw));
+	public function getFootnotes($raw = false) {
+		return ($this->_context->getFootnotes($raw));
 	}
 	/**
 	 * Returns the lines which contain an error.
